Increase the default RAM to 1024 MB for libvirt
The default was 512 MB which was not enough for setting up one host.
The production machine had 1024 MB of memory, so increasing the amount
up to that should be safe.

It would be possible to specify the amount of memory host-by-host in
inventory.yaml but 1024 MB was decided to be a sufficient default amount
for other hosts as well.

Please note this change only affects the libvirt provider.